skills/whatsapp-chatbot-fastapi-starter/assets/fastapi-template/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
